+++ src/components/csp/CategoryMap.tsx
@@ -48,11 +48,15 @@ const MapChart = (props) => {
                                 const AArray = ACur.practiceCategories;
                                 const BCur = statuteRecord.find((s) => s.statuteName === "2014 Eligible Land");
                                 const BArray = BCur.practiceCategories;
-                                const TotalArray = AArray.concat(BArray);
+                                const CCur = statuteRecord.find((s) => s.statuteName === "Miscellaneous Practices");
+                                const CArray = CCur.practiceCategories;
+                                const TotalArray = AArray.concat(BArray).concat(CArray);
                                 if (category === "2018 Practices") {
                                     categoryRecord = statuteRecord[0];
                                 } else if (category === "2014 Eligible Land") {
                                     categoryRecord = statuteRecord[1];
+                                } else if (category === "Miscellaneous Practices") {
+                                    categoryRecord = statuteRecord[2];
                                 } else {
                                     categoryRecord = TotalArray.find((s) => s.practiceCategoryName === category);
                                 }
@@ -182,11 +186,15 @@ const CategoryMap = ({
         const AArray = ACur.practiceCategories;
         const BCur = statuteRecord.find((s) => s.statuteName === "2014 Eligible Land");
         const BArray = BCur.practiceCategories;
-        const TotalArray = AArray.concat(BArray);
+        const CCur = statuteRecord.find((s) => s.statuteName === "Miscellaneous Practices");
+        const CArray = CCur.practiceCategories;
+        const TotalArray = AArray.concat(BArray).concat(CArray);
         if (category === "2018 Practices") {
             categoryRecord = statuteRecord[0];
         } else if (category === "2014 Eligible Land") {
             categoryRecord = statuteRecord[1];
+        } else if (category === "Miscellaneous Practices") {
+            categoryRecord = statuteRecord[2];
         } else {
             categoryRecord = TotalArray.find((s) => s.practiceCategoryName === category);
         }