+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+- Added null value handling for new CSP data [#301](https://github.com/policy-design-lab/pdl-frontend/issues/301) 
+- Added Miscellaneous Practices as a new CSP practice category [#303](https://github.com/policy-design-lab/pdl-frontend/issues/303) 
+
 ## [1.0.2] - 2024-07-23
 
 ### Fixed